Refactor CoroutineScopes and IO in course UI and utils
- Removed manual `CoroutineScope` from `CourseFilterController`, using `viewLifecycleOwner.lifecycleScope` injected from `CoursesFragment`.
- Extracted preview I/O tasks from `InlineResourceAdapter` into a new `ResourcePreviewHelper` injected with `DispatcherProvider`.
- Explicitly canceled `InlineResourceAdapter`'s internal scope in `onDetachedFromRecyclerView`.
- Removed `DefaultDispatcherProvider()` fallbacks from `CameraUtils` and `ANRWatchdog`, enforcing explicit injection.
